Add files via upload
For every organization, today attendance is the most important thing to record the presence of someone. The presence of someone in an organization is a sign that the person is carrying out their obligations to come to the agency or organization.
Usually, attendance is done manually. It can be signed or called one by one. In this digital age, there AQ1 must be a change from this absence to be able to accelerate and provide time efficiency. We can use face recognition to record attendance from everyone present in an organization.
This project will help eliminate the traditional attendance system, minimize manipulation during attendance and record the arrival time of the students. It is also very easy to use and manage. Like every application, there are some setbacks to this application. The application is not one hundred percent accurate. Different factors such as image quality and lack of data sets can decrease the efficiency of the application.
The face recognition based automated student attendance system able to overcome the problem of fraudulent approach and lecturers does not have to count the number of students several times to ensure the presence of the students.
The objective of this project is to develop face recognition attendance system. Expected achievements in order to fulfill the objectives are:

•	To detect the face segment from the video frame.
•	To extract the useful features from the face detected.
•	To classify the features in order to recognize the face detected.
•	To record the attendance of the identified student.

The work on face recognition began in 1960. Woody Bledsoe, Helen Chan Wolf and Charles Bisson had introduced a system which required the administrator to locate eyes, ears, nose and mouth from images. The distance and ratios between the located features and the common reference points are then calculated and compared. The studies are further enhanced by Goldstein, Harmon, and Lesk in 1970 by using other features such as hair colour and lip thickness to automate the recognition. In 1988, Kirby and Sirovich first suggested principle component analysis (PCA) to solve face recognition problem.


A person’s face has distinctive physical shape and characteristics that are used
to identify or verify an individual. Facial recognition records this biometrics of the face. Different face recognition methods measure the biometric of the face.
Facial recognition has become a very important topic in recent years. Facial recognition is effectively applied in various applications like security systems, authentication, entrance control, surveillance system, unlocking of smartphones and social networking systems, etc. Most of the practices do not use facial recognition as the main form of conceding entry. However, with advancement in technology and algorithm, facial recognition system has the potential to replace the standard passwords and fingerprint scanners.

This project was carried out to show how a Local Binary Pattern Histogram (LBPH) face recognizer could be used for taking attendance of students. LBPH facial recognizer is a pre-trained facial recognition classifier. If enough data set are available on the face that is needed to be identified, LBPH can perform facial recognition with high accuracy. Face Recognition Student Attendance System is a desktop application that identifies and verifies student’s identities with the help of a digital image. Once the recognized face matches with the stored image, the attendance is completed and marked in the database for the student.

The facial recognition system has three main phases, which are described below:

1.1	Face Detection
Face detection is the ability to identify the person’s faces within the digital images. This system identifies the human face present in an image or video. We need to define a general structure of a face to determine certain picture or video contains a face (or several). Human faces have the same features such as eyes, nose, forehead, mouth, and chin. Therefore, the objective of face detection is to find the location and size of the face in an image. The located face is then used by the facial recognition algorithm.

1.2	Feature Extraction
In this phase, we are extracting the features from the detected face. In LBPH, the first local binary pattern images are computed, and a histogram is created for facial recognition. This generates a template. A template is a set of data that represents the unique and distinctive features of the detected face.

1.3	Face Recognition 

Face Recognition is being able to uniquely identify and verify a person’s face by comparing and analyzing a biometrics person’s face. A face recognition system is an application that is used for identifying or verifying a person from a digital image.




5.	CONCLUSION 

Automated Attendance System has been envisioned for the purpose of reducing the errors that occur in the traditional (manual) attendance taking system. The aim is to automate and make a system that is useful to the organization such as an institute. The efficient and accurate method of attendance in the office environment that can replace the old manual methods. This method is secure enough, reliable and available for use. No need for specialized hardware for installing the system in the office. It can be constructed using a camera and computer.